% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lol_get_data_champ}
\alias{lol_get_data_champ}
\title{Get information of a champion}
\usage{
lol_get_data_champ(version = "11.1.1", language = "en_US", champion_name)
}
\arguments{
\item{version}{Version of the list of champs, default = 11.1.1. Check https://developer.riotgames.com/docs/lol#data-dragon_champions to get more information}

\item{language}{The language you want to be some character vector such as dialogs. Default is "en_US"}
}
\value{
A data.frame with columns:
}
\description{
This function allows you to get information of a specific champion.
}
\keyword{division,}
\keyword{leagueoflegends}
\keyword{queue,}
\keyword{rank,}
\keyword{region,}
\keyword{summoners,}
