% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lol_get_masteries}
\alias{lol_get_masteries}
\title{Get the champion masteries of a summoner}
\usage{
lol_get_masteries(region = "la1", summoner_id, API)
}
\arguments{
\item{region}{The region where is the summoner you want to query, it should be a character vector with length 1. Valid entries are:
BR1,EUN1    EUW1, JP1, KR, LA1, LA2, NA1, OC1, TR1, RU}

\item{summoner_id}{The summoner id you want to get information about. It must be a character vector with length 1.}

\item{API}{this is your personal API key, to get yours, please, visit the developer portal of Riot Games. The parameter must be a character vector with length 1. This API key expires every 24h.}
}
\value{
A data.frame with columns:
}
\description{
This function allows you to get the data all the champion masteries that a summoner owns based on its account ID (this can be get with lol_get_summoner_data() ).
}
\keyword{division,}
\keyword{leagueoflegends}
\keyword{queue,}
\keyword{rank,}
\keyword{region,}
\keyword{summoners,}
