% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lol_get_summoners_data}
\alias{lol_get_summoners_data}
\title{Get Summoners from a specific region, division, queue, and tier}
\usage{
lol_get_summoners_data(
  region = "la1",
  queue = "RANKED_SOLO_5x5",
  division = "DIAMOND",
  rank = "I",
  API
)
}
\arguments{
\item{region}{The region you want to query, it should be a character vector with length 1. Valid entries are:
BR1,EUN1    EUW1, JP1, KR, LA1, LA2, NA1, OC1, TR1, RU}

\item{queue}{Map or game mode you want to query, it must be a character vector with length 1. Valid entries are:
RANKED_SOLO_5x5, RANKED_TEAM_5x5}

\item{division}{The division you want to query the, it must be a character vector with length 1. Valid entries are:
DIAMOND, PLATINUM. GOLD, SILVER, BRONZE, IRON}

\item{rank}{This is refered to the tier you want to query, it must be a character vector with length 1. Valid entries are:
I, II, III, IV}

\item{API}{this is your personal API key, to get yours, please, visit the developer portal of Riot Games. The parameter must be a character vector with length 1. This API key expires every 24h.}
}
\value{
A data.frame with columns:
}
\description{
This function allows you to get the data from some summoners of the specified region, division, queue and tier.
}
\keyword{division,}
\keyword{leagueoflegends}
\keyword{queue,}
\keyword{rank,}
\keyword{region,}
\keyword{summoners,}
